# enable DeepSeek V3 shared_experts_fusion in sm90