--- a/python/sglang/srt/models/deepseek_v2.py
+++ b/python/sglang/srt/models/deepseek_v2.py
@@ -1383,7 +1383,12 @@ class DeepseekV2ForCausalLM(nn.Module):
                     self.n_share_experts_fusion == self.tp_size
                 ), f"Shared experts fusion optimization is enabled in DeepSeek V3/R1, set it to {self.tp_size} can get best optimized performace."
         elif self.n_share_experts_fusion == 0:
-            if torch.cuda.get_device_capability("cuda") >= (9, 0) and self.config.architectures[0] == "DeepseekV3ForCausalLM" and self.config.n_routed_experts == 256 and (not global_server_args_dict["enable_deepep_moe"]):
+            if (
+                torch.cuda.get_device_capability("cuda") >= (9, 0)
+                and self.config.architectures[0] == "DeepseekV3ForCausalLM"
+                and self.config.n_routed_experts == 256
+                and (not global_server_args_dict["enable_deepep_moe"])
+            ):
                 self.n_share_experts_fusion = self.tp_size
                 global_server_args_dict["n_share_experts_fusion"] = self.tp_size
                 logger.info(